Move physicalUI to /devices. Add bindToState step
